+++ backend/package.json
@@ -26,6 +26,7 @@
     "multer": "^1.4.5-lts.1",
     "multer-storage-cloudinary": "^4.0.0",
     "nodemailer": "^6.9.15",
-    "nodemon": "^3.1.7"
+    "nodemon": "^3.1.7",
+    "sslcommerz-lts": "^1.1.0"
   }
 }
